Particle system and trail shader
This version has addons for both a basic particle system and for the
trail shader, should work with midi input but test using keys 1,2,3 and
4

Time scaling has been implemented.

The image referencing went weird again so the path is absolute.

Move the shaders folder into the data folder of the project (you can
see the reference in trailShaderAddon.cpp line 18)